easyocr/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
opencv-python-headless>=4.5.0
scipy>=1.7.0
numpy>=1.20.0
Pillow>=9.0.0
scikit-image>=0.18.0
python-bidi>=0.4.2
PyYAML>=6.0
Shapely>=1.8.0
pyclipper>=1.3.0
ninja>=1.10.0
easyocr>=1.6.0
openai>=1.0.0
pathlib>=1.0.1
python-dotenv>=0.20.0
flask>=2.0.0
flask-cors>=3.0.10
requests>=2.28.0
gunicorn>=20.1.0
multiprocessing>=2.6.2.1
tqdm>=4.64.0
matplotlib>=3.5.0